+++ spec/cztop/authenticator_spec.rb
@@ -75,4 +75,70 @@ describe CZTop::Authenticator do
       expect(actor).to receive(:<<).with("GSSAPI").and_call_original
     end
   end
+
+  describe "with certificate store" do
+    subject { CZTop::Authenticator.new(cert_store) }
+    let(:cert_store) { CZTop::CertStore.new }
+    let(:cert) { CZTop::Certificate.new }
+    let(:pubkey_z85) { cert.public_key(format: :z85) }
+    let(:pubkey_bin) { cert.public_key(format: :binary) }
+
+    before(:each) do
+      # cache key now, as certificate will be gone later
+      pubkey_z85
+      pubkey_bin
+
+      cert_store.insert(cert)
+    end
+
+    after(:each) do
+      subject.terminate
+    end
+
+    it "initializes" do
+      subject
+    end
+
+    it "uses certificate store passed" do
+      assert_kind_of CZTop::Certificate, cert
+    end
+
+    context "authentication" do
+      let(:domain) { "global" }
+      let(:req) do
+        # REQ socket acting as a CURVE server trying to authenticate a client
+        CZTop::Socket::REQ.new(CZTop::ZAP::ENDPOINT)
+      end
+      rid = 0
+      let(:request_id) { "#{rid += 1}" }
+      let(:zap_request) do
+        CZTop::ZAP::Request.new(domain, credentials).tap do |r|
+          r.request_id = request_id
+        end
+      end
+
+      let(:zap_response) { CZTop::ZAP::Response.from_message(req.receive) }
+
+      before(:each) do
+        subject # start authenticator
+        req << zap_request.to_msg
+      end
+
+      context "with valid credentials" do
+        let(:credentials) { [pubkey_bin] }
+        it "authenticates" do
+          assert_operator zap_response, :success?
+          assert_equal request_id, zap_response.request_id
+        end
+      end
+      context "with invalid credentials" do
+        let(:credentials) { ["f" * 32] } # unknown public key
+        it "does not authenticate" do
+          refute_operator zap_response, :success?
+          assert_equal request_id, zap_response.request_id
+          assert_nil zap_response.user_id
+        end
+      end
+    end
+  end
 end